+++ Sources/ParseSwift/Objects/ParseInstallation.swift
@@ -97,7 +97,7 @@ public extension ParseInstallation {
     }
 
     func mergeParse(with object: Self) throws -> Self {
-        guard hasSameObjectId(as: object) == true else {
+        guard hasSameObjectId(as: object) else {
             throw ParseError(code: .unknownError,
                              message: "objectId's of objects do not match")
         }
@@ -193,7 +193,7 @@ extension ParseInstallation {
 }
 
 // MARK: CurrentInstallationContainer
-struct CurrentInstallationContainer<T: ParseInstallation>: Codable {
+struct CurrentInstallationContainer<T: ParseInstallation>: Codable, Hashable {
     var currentInstallation: T?
     var installationId: String?
 }