+++ Tests/ParseSwiftTests/KeychainStoreTests.swift
@@ -5,6 +5,7 @@
 //  Created by Florent Vilmart on 17-09-25.
 //  Copyright © 2020 Parse Community. All rights reserved.
 //
+
 #if !os(Linux) && !os(Android) && !os(Windows)
 import Foundation
 import XCTest
@@ -14,12 +15,24 @@ class KeychainStoreTests: XCTestCase {
     var testStore: KeychainStore!
     override func setUpWithError() throws {
         try super.setUpWithError()
+        guard let url = URL(string: "http://localhost:1337/1") else {
+            XCTFail("Should create valid URL")
+            return
+        }
+        ParseSwift.initialize(applicationId: "applicationId",
+                              clientKey: "clientKey",
+                              masterKey: "masterKey",
+                              serverURL: url, testing: true)
         testStore = KeychainStore(service: "test")
     }
 
     override func tearDownWithError() throws {
         try super.tearDownWithError()
         _ = testStore.removeAllObjects()
+        #if !os(Linux) && !os(Android) && !os(Windows)
+        try KeychainStore.shared.deleteAll()
+        #endif
+        try ParseStorage.shared.deleteAll()
     }
 
     func testSetObject() {
@@ -167,6 +180,79 @@ class KeychainStoreTests: XCTestCase {
             XCTAssertTrue(testStore.set(object: "yarr", forKey: "pirate2"), "Should set value")
             XCTAssertTrue(testStore.removeAllObjects(), "Should set value")
         }
+    }
+
+    func testQueryTemplate() throws {
+        let query = KeychainStore.shared.getKeychainQueryTemplate()
+        XCTAssertEqual(query.count, 2)
+        XCTAssertEqual(query[kSecAttrService as String] as? String, KeychainStore.shared.service)
+        XCTAssertEqual(query[kSecClass as String] as? String, kSecClassGenericPassword as String)
+    }
+
+    func testQueryNoAccessGroup() throws {
+        let accessGroup = ParseKeychainAccessGroup()
+        let query = KeychainStore.shared.keychainQuery(forKey: "hello", accessGroup: accessGroup)
+        XCTAssertEqual(query.count, 5)
+        XCTAssertEqual(query[kSecAttrService as String] as? String, KeychainStore.shared.service)
+        XCTAssertEqual(query[kSecClass as String] as? String, kSecClassGenericPassword as String)
+        XCTAssertEqual(query[kSecAttrAccount as String] as? String, "hello")
+        XCTAssertEqual(query[kSecAttrSynchronizable as String] as? Bool, kCFBooleanFalse as? Bool)
+        XCTAssertEqual(query[kSecAttrAccessible as String] as? String,
+                       kSecAttrAccessibleAfterFirstUnlockThisDeviceOnly as String)
+    }
+
+    func testQueryAccessGroupSyncableKeyTrue() throws {
+        let accessGroup = ParseKeychainAccessGroup(accessGroup: "world", isSyncingKeychainAcrossDevices: true)
+        let query = KeychainStore.shared.keychainQuery(forKey: "hello", accessGroup: accessGroup)
+        XCTAssertEqual(query.count, 6)
+        XCTAssertEqual(query[kSecAttrService as String] as? String, KeychainStore.shared.service)
+        XCTAssertEqual(query[kSecClass as String] as? String, kSecClassGenericPassword as String)
+        XCTAssertEqual(query[kSecAttrAccount as String] as? String, "hello")
+        XCTAssertEqual(query[kSecAttrAccessGroup as String] as? String, "world")
+        XCTAssertEqual(query[kSecAttrSynchronizable as String] as? Bool, kCFBooleanTrue as? Bool)
+        XCTAssertEqual(query[kSecAttrAccessible as String] as? String,
+                       kSecAttrAccessibleAfterFirstUnlock as String)
+    }
+
+    func testQueryAccessGroupSyncableKeyFalse() throws {
+        let accessGroup = ParseKeychainAccessGroup(accessGroup: "world", isSyncingKeychainAcrossDevices: false)
+        let query = KeychainStore.shared.keychainQuery(forKey: "hello", accessGroup: accessGroup)
+        XCTAssertEqual(query.count, 6)
+        XCTAssertEqual(query[kSecAttrService as String] as? String, KeychainStore.shared.service)
+        XCTAssertEqual(query[kSecClass as String] as? String, kSecClassGenericPassword as String)
+        XCTAssertEqual(query[kSecAttrAccount as String] as? String, "hello")
+        XCTAssertEqual(query[kSecAttrAccessGroup as String] as? String, "world")
+        XCTAssertEqual(query[kSecAttrSynchronizable as String] as? Bool, kCFBooleanFalse as? Bool)
+        XCTAssertEqual(query[kSecAttrAccessible as String] as? String,
+                       kSecAttrAccessibleAfterFirstUnlockThisDeviceOnly as String)
+    }
+
+    func testQueryAccessGroupNoSyncableKeyTrue() throws {
+        let key = ParseStorage.Keys.currentInstallation
+        let accessGroup = ParseKeychainAccessGroup(accessGroup: "world", isSyncingKeychainAcrossDevices: true)
+        let query = KeychainStore.shared.keychainQuery(forKey: key, accessGroup: accessGroup)
+        XCTAssertEqual(query.count, 6)
+        XCTAssertEqual(query[kSecAttrService as String] as? String, KeychainStore.shared.service)
+        XCTAssertEqual(query[kSecClass as String] as? String, kSecClassGenericPassword as String)
+        XCTAssertEqual(query[kSecAttrAccount as String] as? String, key)
+        XCTAssertEqual(query[kSecAttrAccessGroup as String] as? String, "world")
+        XCTAssertEqual(query[kSecAttrSynchronizable as String] as? Bool, kCFBooleanFalse as? Bool)
+        XCTAssertEqual(query[kSecAttrAccessible as String] as? String,
+                       kSecAttrAccessibleAfterFirstUnlockThisDeviceOnly as String)
+    }
+
+    func testQueryAccessGroupNoSyncableKeyFalse() throws {
+        let key = ParseStorage.Keys.currentInstallation
+        let accessGroup = ParseKeychainAccessGroup(accessGroup: "world", isSyncingKeychainAcrossDevices: false)
+        let query = KeychainStore.shared.keychainQuery(forKey: key, accessGroup: accessGroup)
+        XCTAssertEqual(query.count, 6)
+        XCTAssertEqual(query[kSecAttrService as String] as? String, KeychainStore.shared.service)
+        XCTAssertEqual(query[kSecClass as String] as? String, kSecClassGenericPassword as String)
+        XCTAssertEqual(query[kSecAttrAccount as String] as? String, key)
+        XCTAssertEqual(query[kSecAttrAccessGroup as String] as? String, "world")
+        XCTAssertEqual(query[kSecAttrSynchronizable as String] as? Bool, kCFBooleanFalse as? Bool)
+        XCTAssertEqual(query[kSecAttrAccessible as String] as? String,
+                       kSecAttrAccessibleAfterFirstUnlockThisDeviceOnly as String)
     }
 }
 #endif